Assembler built
The assembler is built and (almost) finished. It passes all the test except for Pong.asm, however, for class, all I have to do is pass Rect.asm, so I'm calling it good. If I had more time (and patience) I would double-check what's wrong and finish this. But since the program is so big to read line by line of assembly, I'll leave it be for now.